sdk: go: move into own module
