+++ codegen/generator/templates/src/header.go.tmpl
@@ -6,7 +6,7 @@ import (
 	"context"
 
 	"github.com/Khan/genqlient/graphql"
-	"github.com/dagger/dagger/sdk/go/dagger/querybuilder"
+	"dagger.io/dagger/querybuilder"
 )
 
 // New returns a new API query object
