+++ core/integration/container_test.go
@@ -5,11 +5,11 @@ import (
 	"strings"
 	"testing"
 
+	"dagger.io/dagger"
+	"dagger.io/dagger/api"
 	"github.com/dagger/dagger/core"
 	"github.com/dagger/dagger/core/schema"
 	"github.com/dagger/dagger/internal/testutil"
-	"github.com/dagger/dagger/sdk/go/dagger"
-	"github.com/dagger/dagger/sdk/go/dagger/api"
 	"github.com/moby/buildkit/identity"
 	"github.com/stretchr/testify/require"
 )
