+++ core/integration/examples_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/dagger/dagger/sdk/go/dagger"
+	"dagger.io/dagger"
 	"github.com/stretchr/testify/require"
 )
