+++ core/integration/project_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/dagger/dagger/sdk/go/dagger"
+	"dagger.io/dagger"
 	"github.com/stretchr/testify/require"
 )
