+++ sdk/go/api/.gitattributes
