+++ sdk/go/api/api.gen.go
@@ -5,8 +5,8 @@ package api
 import (
 	"context"
 
+	"dagger.io/dagger/querybuilder"
 	"github.com/Khan/genqlient/graphql"
-	"github.com/dagger/dagger/sdk/go/dagger/querybuilder"
 )
 
 // New returns a new API query object
