+++ sdk/go/server.go
@@ -14,7 +14,7 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/dagger/dagger/sdk/go/dagger/api"
+	"dagger.io/dagger/api"
 	"github.com/iancoleman/strcase"
 	"github.com/vektah/gqlparser/v2/ast"
 	"github.com/vektah/gqlparser/v2/formatter"
@@ -148,7 +148,7 @@ func (ts *goTypes) schema() []byte {
 		The much cleaner approach will come when we integrate this code w/ the
 		in-progress codegen work.
 		*/
-		if strings.HasPrefix(s.typ.PkgPath(), "github.com/dagger/dagger/sdk/go") {
+		if strings.HasPrefix(s.typ.PkgPath(), "dagger.io/dagger") {
 			continue
 		}
 
@@ -504,7 +504,7 @@ func goReflectTypeToGraphqlType(t reflect.Type, isInput bool) *ast.Type {
 		The much cleaner approach will come when we integrate this code w/ the
 		in-progress codegen work.
 		*/
-		if strings.HasPrefix(t.PkgPath(), "github.com/dagger/dagger/sdk/go") {
+		if strings.HasPrefix(t.PkgPath(), "dagger.io/dagger") {
 			return ast.NonNullNamedType(t.Name(), nil)
 		}
 		return ast.NonNullNamedType("String", nil)
